restructured project to separate frontend and backend
