+++ api/.gitignore
@@ -2,5 +2,5 @@
 /node_modules
 package-lock.json
 
-# environment variables
-.env
+# build
+/dist
