+++ api/package.json
@@ -1,0 +1,25 @@
+{
+	"name": "songselect-api",
+	"version": "0.1.0",
+	"main": "server.ts",
+	"scripts": {
+		"start": "concurrently \"tsc -w\" \"nodemon ./dist/server.js\"",
+		"build": "tsc",
+		"start-client": "cd ../client && npm run start"
+	},
+	"dependencies": {
+		"@types/body-parser": "^1.19.0",
+		"@types/cors": "^2.8.10",
+		"@types/dotenv": "^8.2.0",
+		"@types/express": "^4.17.12",
+		"body-parser": "^1.19.0",
+		"cors": "^2.8.5",
+		"dotenv": "^10.0.0",
+		"express": "^4.17.1",
+		"typescript": "^4.3.2"
+	},
+	"devDependencies": {
+		"nodemon": "^2.0.7"
+	},
+	"type": "module"
+}
