+++ client/package.json
@@ -1,0 +1,50 @@
+{
+	"name": "songselect-client",
+	"version": "0.1.0",
+	"private": true,
+	"proxy": "http://localhost:8080",
+	"dependencies": {
+		"@reduxjs/toolkit": "^1.5.1",
+		"@testing-library/jest-dom": "^4.2.4",
+		"@testing-library/react": "^9.5.0",
+		"@testing-library/user-event": "^7.2.1",
+		"@types/cors": "^2.8.10",
+		"@types/express": "^4.17.12",
+		"@types/jest": "^24.9.1",
+		"@types/node": "^12.20.13",
+		"@types/react": "^16.14.6",
+		"@types/react-dom": "^16.9.13",
+		"@types/react-redux": "^7.1.16",
+		"body-parser": "^1.19.0",
+		"cors": "^2.8.5",
+		"dotenv": "^8.2.0",
+		"express": "^4.17.1",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2",
+		"react-redux": "^7.2.4",
+		"react-scripts": "4.0.3",
+		"typescript": "^4.1.5"
+	},
+	"scripts": {
+		"start": "react-scripts start",
+		"build": "react-scripts build",
+		"test": "react-scripts test",
+		"eject": "react-scripts eject",
+		"start-server": "cd ../api && npm run start"
+	},
+	"eslintConfig": {
+		"extends": "react-app"
+	},
+	"browserslist": {
+		"production": [
+			">0.2%",
+			"not dead",
+			"not op_mini all"
+		],
+		"development": [
+			"last 1 chrome version",
+			"last 1 firefox version",
+			"last 1 safari version"
+		]
+	}
+}
