+++ client/src/App.tsx
@@ -1,0 +1,70 @@
+import React, { useEffect, useState } from 'react';
+import logo from './logo.svg';
+import { Counter } from './features/counter/Counter';
+import './App.css';
+
+function App() {
+	const [fromAPI, setFromAPI] = useState('');
+
+	const callAPI = async () => {
+		const response: Response = await fetch('/ping');
+		const message: string = await response.text();
+		setFromAPI(message);
+	}
+
+	useEffect(() => {
+		callAPI();
+	});
+
+	return (
+		<div className="App">
+		<header className="App-header">
+			<img src={logo} className="App-logo" alt="logo" />
+			<Counter />
+			<p>
+			{`ExpressJS API Test: ${fromAPI}`}
+			</p>
+			<span>
+			<span>Learn </span>
+			<a
+				className="App-link"
+				href="https://reactjs.org/"
+				target="_blank"
+				rel="noopener noreferrer"
+			>
+				React
+			</a>
+			<span>, </span>
+			<a
+				className="App-link"
+				href="https://redux.js.org/"
+				target="_blank"
+				rel="noopener noreferrer"
+			>
+				Redux
+			</a>
+			<span>, </span>
+			<a
+				className="App-link"
+				href="https://redux-toolkit.js.org/"
+				target="_blank"
+				rel="noopener noreferrer"
+			>
+				Redux Toolkit
+			</a>
+			,<span> and </span>
+			<a
+				className="App-link"
+				href="https://react-redux.js.org/"
+				target="_blank"
+				rel="noopener noreferrer"
+			>
+				React Redux
+			</a>
+			</span>
+		</header>
+		</div>
+	);
+}
+
+export default App;
