+++ package.json
@@ -1,42 +1,23 @@
 {
-  "name": "songselect",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@reduxjs/toolkit": "^1.5.1",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/jest": "^24.9.1",
-    "@types/node": "^12.20.13",
-    "@types/react": "^16.14.6",
-    "@types/react-dom": "^16.9.13",
-    "@types/react-redux": "^7.1.16",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.1.5"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
+	"name": "songselect",
+	"version": "0.1.0",
+	"description": "Hello, this is the beginning of our song select game that will be programmed using Node.js, React-Redux, and TypeScript.",
+	"scripts": {
+		"start-dev": "concurrently \"cd ./api && npm run start\" \"cd ./client && npm run start\"",
+		"install-dev": "npm install && concurrently \"cd ./api && npm install --production=false\" \"cd ./client && npm install --production=false\"",
+		"build": "concurrently \"cd ./api && npm run build\" \"cd ./client && npm run build\""
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/simonlee9504/songselect.git"
+	},
+	"author": "Tony Tian, Simon Lee, and Ethan Yap",
+	"license": "ISC",
+	"bugs": {
+		"url": "https://github.com/simonlee9504/songselect/issues"
+	},
+	"homepage": "https://github.com/simonlee9504/songselect#readme",
+	"devDependencies": {
+		"concurrently": "^6.2.0"
+	}
 }
